+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ensure .NET Framework builds pass PEVerify (#744)
 - `OpenXmlPartContainer.DeletePart` no longer throws an exception if there isn't a match for the identifier given (#740)
 - Mark obsolete members to not show up with Intellisense (#745)
+- Fixed issue with `AttributeRequiredConditionToValue` semantic constraint where validation could fail on correct input (#746)
 
 ## Version 2.11.0 - 2020-05-21
 ### Added